Put the 'home' directory under version control
This will allow a user to easily sync with nanwan's home directory if we
add stuff in there in the future. By default we .gitignore *everything*
so adding files in the home directory will not pollute the git repo.  As
nanwan we will allways 'git add . -f', forcing our files into the repo
for the user to sync with.